agents/ranker/requirements.txt
# Ranker Agent Requirements
python-dotenv>=1.0.0
nats-py>=2.4.0
asyncpg>=0.28.0
pydantic>=2.4.2
asyncio>=3.4.3
tenacity>=8.0.1
openai>=1.3.0
pgvector>=0.2.3
scikit-learn>=1.3.2
numpy>=1.26.0
python-dateutil>=2.8.2